Revert "feat(mobile): time relay dial stages so diagnostics say where a slow connect went (#19245)"
This reverts commit 83b1558ecce27decfe46fe01de8b95a4b4d1337e.
