+++ mobile/src/diagnostics/connection-diagnostics-report.test.ts
@@ -1,35 +1,7 @@
 import { describe, expect, it } from 'vitest'
 import { buildConnectionDiagnosticsReport } from './connection-diagnostics-report'
-import type { ConnectionLogEntry } from '../transport/types'
 
 const NOW = Date.UTC(2026, 6, 9, 22, 0, 0)
-
-function stageEntry(
-  id: string,
-  ts: number,
-  name: string,
-  ms: number,
-  complete: boolean
-): ConnectionLogEntry {
-  return {
-    id,
-    ts,
-    level: complete ? 'info' : 'warn',
-    path: 'relay',
-    message: `Relay dial stage ${name} ${complete ? 'finished' : 'did not finish'}`,
-    timing: { kind: 'relay-dial-stage', name, ms, complete }
-  }
-}
-
-function stateEntry(id: string, ts: number, name: string, ms: number): ConnectionLogEntry {
-  return {
-    id,
-    ts,
-    level: 'info',
-    message: `Connection state ${name} → connected`,
-    timing: { kind: 'connection-state', name, ms, complete: true }
-  }
-}
 
 describe('buildConnectionDiagnosticsReport', () => {
   it('summarizes a failing Tailscale host with its log', () => {
@@ -97,27 +69,12 @@ describe('buildConnectionDiagnosticsReport', () => {
       pendingPath: 'relay',
       entries: [
         {
-          id: 'relay-stage-opening',
-          ts: NOW - 6_000,
-          level: 'info',
-          path: 'relay',
-          message: 'Relay dial stage opening finished',
-          detail: '118ms — resumeToken=secret-resume-token',
-          timing: { kind: 'relay-dial-stage', name: 'opening', ms: 118, complete: true }
-        },
-        {
           id: 'relay-failure',
           ts: NOW - 5_000,
           level: 'error',
           message: 'Relay: relay dial failed',
           detail:
-            'RelayDirectorHttpError: relay director resolve failed (503); retry after 30000ms; resumeToken=secret-resume-token',
-          timing: {
-            kind: 'relay-dial-stage',
-            name: 'awaiting-hello',
-            ms: 9_100,
-            complete: false
-          }
+            'RelayDirectorHttpError: relay director resolve failed (503); retry after 30000ms; resumeToken=secret-resume-token'
         }
       ],
       nowMs: NOW
@@ -130,9 +87,6 @@ describe('buildConnectionDiagnosticsReport', () => {
     expect(report).toContain('Next step: Keep Orca open; recovery should retry automatically.')
     expect(report).toContain('resumeToken=[redacted]')
     expect(report).not.toContain('secret-resume-token')
-    expect(report).toContain(
-      'Relay dial stages: opening 118ms · awaiting-hello 9.1s (did not finish) — total 9.2s'
-    )
   })
 
   it('redacts quoted JSON credentials and never echoes an invalid endpoint', () => {
@@ -160,52 +114,6 @@ describe('buildConnectionDiagnosticsReport', () => {
     expect(report).not.toContain('endpoint-secret')
     expect(report).not.toContain('json-secret')
     expect(report).not.toContain('bearer-secret')
-  })
-
-  it('breaks a slow connect down by dial stage and connection state', () => {
-    const report = buildConnectionDiagnosticsReport({
-      hostName: 'Host 6',
-      endpoint: 'ws://192.168.1.50:6768',
-      state: 'connected',
-      reconnectAttempts: 2,
-      lastConnectedAt: NOW,
-      platform: 'ios 26.5.1',
-      appVersion: '0.0.47',
-      entries: [
-        stageEntry('a1', NOW - 30_000, 'opening', 90, false),
-        stateEntry('s1', NOW - 29_000, 'connecting', 12_000),
-        stageEntry('b1', NOW - 20_000, 'opening', 120, true),
-        stageEntry('b2', NOW - 19_000, 'awaiting-hello', 6_400, true),
-        stageEntry('b3', NOW - 13_000, 'handshaking', 240, true),
-        stageEntry('b4', NOW - 12_000, 'confirming', 1_180, true),
-        stateEntry('s2', NOW - 11_000, 'connecting', 8_000)
-      ],
-      nowMs: NOW
-    })
-
-    // Only the latest dial is broken out, so a reconnect loop cannot average away
-    // the attempt the reporter is complaining about.
-    expect(report).toContain(
-      'Relay dial stages (latest of 2): opening 120ms · awaiting-hello 6.4s · handshaking 240ms · confirming 1.2s — total 7.9s'
-    )
-    expect(report).toContain('Connection state dwell: connecting 20.0s ×2')
-  })
-
-  it('omits the timing lines when nothing recorded a phase duration', () => {
-    const report = buildConnectionDiagnosticsReport({
-      hostName: 'Host 7',
-      endpoint: 'ws://192.168.1.50:6768',
-      state: 'connected',
-      reconnectAttempts: 0,
-      lastConnectedAt: NOW,
-      platform: 'ios 26.5.1',
-      appVersion: '0.0.47',
-      entries: [{ id: 'plain', ts: NOW, level: 'info', message: 'Authenticated' }],
-      nowMs: NOW
-    })
-
-    expect(report).not.toContain('Relay dial stages')
-    expect(report).not.toContain('Connection state dwell')
   })
 
   it('bounds a single event line before submission while preserving its identity', () => {
