+++ mobile/src/transport/direct-connection-log.ts
@@ -4,14 +4,8 @@ import type {
   ConnectionLogEntry,
   ConnectionLogLevel,
   ConnectionLogSink,
-  ConnectionState,
   MobileConnectionDiagnosticPath
 } from './types'
-
-// Why: every reconnect cycle walks four states, and the per-host buffer is capped.
-// Logging sub-100ms transitions would halve the history a report can show while
-// telling support nothing — those states are never where a slow connect spent time.
-const MIN_LOGGED_DWELL_MS = 100
 
 export class DirectConnectionLog {
   private sequence = 0
@@ -28,7 +22,7 @@ export class DirectConnectionLog {
     level: ConnectionLogLevel,
     message: string,
     detail?: string,
-    evidence?: Pick<ConnectionLogEntry, 'code' | 'path' | 'timing'>
+    evidence?: Pick<ConnectionLogEntry, 'code' | 'path'>
   ): void => {
     this.sink?.({
       id: `log-${++this.sequence}-${Date.now()}`,
@@ -48,26 +42,6 @@ export class DirectConnectionLog {
       `${evidence.reason}; ${evidence.missedProbes}/${evidence.missedProbeLimit} probes missed; last authenticated activity ${evidence.lastInboundAgeMs}ms ago`,
       { code: 'liveness-timeout' }
     )
-  }
-
-  // Why: how long the client sat in each ConnectionState used to go only to
-  // console, so a shared diagnostics report could not show where a slow connect
-  // spent its seconds.
-  stateDwell = (previous: ConnectionState, next: ConnectionState, dweltMs: number): void => {
-    if (dweltMs < MIN_LOGGED_DWELL_MS) {
-      return
-    }
-    this.emit('info', `Connection state ${previous} → ${next}`, `${dweltMs}ms in ${previous}`, {
-      timing: { kind: 'connection-state', name: previous, ms: dweltMs, complete: true }
-    })
-  }
-
-  retryScheduled = (message: string, detail?: string): void => {
-    this.emit('info', message, detail, { code: 'retry-scheduled' })
-  }
-
-  authenticationRejected = (message: string, detail?: string): void => {
-    this.emit('warn', message, detail, { code: 'authentication-rejected' })
   }
 
   connected = (): void => {
