+++ mobile/src/transport/mobile-relay-rpc-session.ts
@@ -9,18 +9,25 @@ import { MobileE2EEAuthenticationError } from './mobile-e2ee-v2-physical-channel
 import { markRpcDeliveryUnknown } from './rpc-delivery-ambiguity'
 import { openRpcRequestBudget, resolvePostConnectRequestTimeout } from './rpc-request-budget'
 import { isRpcResponse } from './rpc-response-shape'
-import { RelayDialStageLog } from './relay-dial-stage-log'
 import { RelayDialStageTracker, type RelayDialStageSource } from './relay-dial-stage'
 import { RelayPendingRequests } from './relay-pending-requests'
-import { createRelaySessionLivenessWatchdog } from './relay-session-liveness-profile'
+import { RpcSessionLivenessWatchdog } from './rpc-session-liveness-watchdog'
 import { settleMobileRuntimeCapabilities } from './mobile-runtime-capability-negotiation'
 import type { RelayHostCloseReason } from '../../../src/shared/relay-host-close-reason'
 import type { RpcClient } from './rpc-client'
 import type { ConnectionLogSink, ConnectionState, RpcResponse } from './types'
 
+// Ordinary foreground checks: two 4s misses, at most one voluntary probe per 10s.
+const RELAY_PROBE = { timeoutMs: 4_000, missedProbeLimit: 2, minIntervalMs: 10_000 }
+// A socket that died while the process was suspended must be admitted before the
+// user reads the screen as broken. Two 2s misses, not one: the first frame after a
+// resume rides a cold radio, and a single slow answer is not proof of a dead link.
+const RELAY_RESUME_PROBE = { timeoutMs: 2_000, missedProbeLimit: 2 }
 // Bounds the confirm exactly as migrateTo's own wait used to, so the supervisor's
 // mutex is never held for the full request timeout waiting on a silent cell.
 const RELAY_CONFIRM_TIMEOUT_MS = 12_000
+// Foreground-only sweep so a silently-dead relay surfaces without a user action.
+const RELAY_IDLE_PROBE_MS = 25_000
 let relayRpcSessionSequence = 0
 
 export type MobileRelayRpcSession = RpcClient &
@@ -73,7 +80,6 @@ export function connectMobileRelayRpcSession(args: {
     settleResumeConfirmed = resolve
   })
   const dialStage = new RelayDialStageTracker()
-  const dialStageLog = new RelayDialStageLog(dialStage, logSessionId, args.onLog)
   const streams = new MobileRelayRpcStreams({
     nextId: () => pending.nextId(),
     sendFrame,
@@ -88,7 +94,7 @@ export function connectMobileRelayRpcSession(args: {
     desktopPublicKeyB64: args.desktopPublicKeyB64,
     createSocket: args.createSocket,
     onHostCloseReason: args.onHostCloseReason,
-    onOpen: () => dialStageLog.enter('awaiting-hello'),
+    onOpen: () => dialStage.advance('awaiting-hello'),
     onHello: (hello) => {
       if (
         hello.credentialKind !== 'resume' ||
@@ -99,7 +105,7 @@ export function connectMobileRelayRpcSession(args: {
       }
       attachDeadlineAt = hello.leaseExpiresAt
       resumeExpiresAt = hello.resumeExpiresAt
-      dialStageLog.enter('handshaking')
+      dialStage.advance('handshaking')
       publishState('handshaking')
     },
     onAuthenticated: () => publishAuthenticated(),
@@ -153,14 +159,30 @@ export function connectMobileRelayRpcSession(args: {
     whenResumeConfirmed: () => resumeConfirmed,
     getFailure: () => failure
   }
-  const livenessWatchdog = createRelaySessionLivenessWatchdog({
-    isForeground: args.isForeground,
+  const livenessWatchdog = new RpcSessionLivenessWatchdog({
+    transport: 'relay',
+    idleProbeMs: RELAY_IDLE_PROBE_MS,
+    probeTimeoutMs: RELAY_PROBE.timeoutMs,
+    missedProbeLimit: RELAY_PROBE.missedProbeLimit,
+    voluntaryProbeMinIntervalMs: RELAY_PROBE.minIntervalMs,
+    urgentProbeTimeoutMs: RELAY_RESUME_PROBE.timeoutMs,
+    urgentMissedProbeLimit: RELAY_RESUME_PROBE.missedProbeLimit,
+    shouldIdleProbe: () => args.isForeground?.() ?? true,
     sendProbe: () =>
       state === 'connected' &&
       sendFrame({ id: pending.nextId(), method: 'status.get', params: undefined }),
-    terminate: () => fail(new Error('relay session liveness timeout')),
-    onLog: args.onLog,
-    nextLogId: () => `relay-liveness-${logSessionId}-${++logSequence}`
+    onTimeout: (evidence) => {
+      args.onLog?.({
+        id: `relay-liveness-${logSessionId}-${++logSequence}`,
+        ts: Date.now(),
+        level: 'error',
+        code: 'liveness-timeout',
+        path: 'relay',
+        message: 'Relay health check failed',
+        detail: `${evidence.reason}; ${evidence.missedProbes}/${evidence.missedProbeLimit} probes missed; last authenticated activity ${evidence.lastInboundAgeMs}ms ago`
+      })
+    },
+    terminate: () => fail(new Error('relay session liveness timeout'))
   })
   return client
 
@@ -171,7 +193,7 @@ export function connectMobileRelayRpcSession(args: {
     if (closed) {
       return
     }
-    dialStageLog.enter('confirming')
+    dialStage.advance('confirming')
     void confirmResume().then(settleResumeConfirmed, settleResumeConfirmed)
     // Why: an unanswered advisory says nothing, but a frame that never reached the
     // wire proves the socket cannot carry traffic — that alone still fails.
@@ -202,9 +224,6 @@ export function connectMobileRelayRpcSession(args: {
       }
       resumeConfirmation = result.resumeConfirmation
       resumeExpiresAt = result.resumeConfirmation.resumeExpiresAt
-      // The dial's last stage ends when the desktop has confirmed the resume, not when
-      // 'connected' was published at authentication ahead of it.
-      dialStageLog.settle(true)
     } catch (error) {
       fail(asError(error))
     }
@@ -306,7 +325,6 @@ export function connectMobileRelayRpcSession(args: {
     }
     closed = true
     settleResumeConfirmed()
-    dialStageLog.settle(false, error.message)
     livenessWatchdog.stop(livenessIdentity)
     streams.clear()
     link.close()
