+++ mobile/src/transport/types.ts
@@ -58,17 +58,6 @@ export type ConnectionDiagnosticCode =
   | 'relay-credential-unavailable'
   | 'host-open-failed'
 
-// Why: a 10s connect used to read as one opaque "connecting" span. Attaching the
-// duration of the phase an entry closes out lets the report say where the time
-// went. Diagnostics only — nothing schedules from these.
-export type ConnectionLogTiming = {
-  kind: 'relay-dial-stage' | 'connection-state'
-  name: string
-  ms: number
-  // False when the phase never finished (the dial died inside it).
-  complete: boolean
-}
-
 export type ConnectionLogEntry = {
   id: string
   ts: number
@@ -79,7 +68,6 @@ export type ConnectionLogEntry = {
   detail?: string
   code?: ConnectionDiagnosticCode
   path?: MobileConnectionDiagnosticPath
-  timing?: ConnectionLogTiming
 }
 
 export type ConnectionLogSink = (entry: ConnectionLogEntry) => void
@@ -88,7 +76,7 @@ export type ConnectionLogEmitter = (
   level: ConnectionLogLevel,
   message: string,
   detail?: string,
-  evidence?: Pick<ConnectionLogEntry, 'code' | 'path' | 'timing'>
+  evidence?: Pick<ConnectionLogEntry, 'code' | 'path'>
 ) => void
 
 export type ConnectionState =
@@ -98,16 +86,6 @@ export type ConnectionState =
   | 'disconnected'
   | 'reconnecting'
   | 'auth-failed'
-
-// Exhaustive by construction; see RELAY_DIAL_STAGE_NAMES for why.
-export const CONNECTION_STATE_NAMES: Record<ConnectionState, true> = {
-  connecting: true,
-  handshaking: true,
-  connected: true,
-  disconnected: true,
-  reconnecting: true,
-  'auth-failed': true
-}
 
 // Why: a user-attention nudge must not tear down a healthy relay (probe it); only a
 // network-change nudge marks the socket suspect enough to replace it.
